ecommerce/services/cart_service/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
psycopg2-binary==2.9.9
PyJWT==2.9.0
requests==2.32.3
gunicorn==22.0.0
